+++ utils/convertions.py
@@ -8,3 +8,12 @@ def convert_decimal(doc):
         if type(value) == Decimal:
             doc[key] = Decimal128(value)
     return doc
+
+
+def convert_decimal_for_response(document):
+    for key, value in document.items():
+        if type(value) == Decimal:
+            document[key] = float(value)
+        elif type(value) == Decimal128:
+            document[key] = float(str(value))
+    return document
